+++ tests/test.nim
@@ -24,3 +24,4 @@ import ./t_item_processing
 import ./t_lockfree_stack_typestates
 import ./t_backoff
 import ./t_bind_client
+import ./t_manager_cc_surface